move the settings class
